Move xml_utils into drake/multibody/parsers.
