+++ drake/multibody/parsers/sdf_parser.cc
@@ -16,8 +16,8 @@
 #include "drake/multibody/joints/floating_base_types.h"
 #include "drake/multibody/parsers/model_instance_id_table.h"
 #include "drake/multibody/parsers/parser_common.h"
+#include "drake/multibody/parsers/xml_util.h"
 #include "drake/multibody/rigid_body_tree.h"
-#include "drake/multibody/xml_util.h"
 #include "drake/thirdParty/zlib/tinyxml2/tinyxml2.h"
 
 // from
