+++ drake/multibody/parsers/test/CMakeLists.txt
@@ -1,2 +1,19 @@
+add_executable(urdf_kinematics_test urdf_kinematics_test.cc)
+target_link_libraries(urdf_kinematics_test
+    drakeParsers
+    drakeRBM)
+if (Bullet_FOUND)
+  add_executable(urdf_collision_test urdf_collision_test.cc)
+  target_link_libraries(urdf_collision_test drakeParsers drakeRBM)
+endif ()
+
+add_executable(urdf_manipulator_dynamics_test urdf_manipulator_dynamics_test.cc)
+target_link_libraries(urdf_manipulator_dynamics_test
+    drakeParsers
+    drakeRBM)
+
+drake_add_cc_test(xml_util_test)
+target_link_libraries(xml_util_test drakeParsers)
+
 add_subdirectory(package_map_test)
 add_subdirectory(urdf_parser_test)
