+++ drake/multibody/parsers/test/urdf_collision_test.cc
