+++ drake/multibody/parsers/test/urdf_kinematics_test.cc
