+++ drake/multibody/parsers/test/urdf_manipulator_dynamics_test.cc
