+++ drake/multibody/parsers/urdf_parser.h
@@ -10,7 +10,7 @@
 #include "drake/multibody/parsers/model_instance_id_table.h"
 #include "drake/multibody/parsers/package_map.h"
 #include "drake/multibody/parsers/parser_common.h"
-#include "drake/multibody/xml_util.h"
+#include "drake/multibody/parsers/xml_util.h"
 #include "drake/thirdParty/zlib/tinyxml2/tinyxml2.h"
 
 namespace drake {
